// src/types/productsPage.enums.ts
export enum ProductsPageTexts {
  primaryHeader = "Swag Labs",
  title = "Products",
}

export enum ProductsImages {
  SauceLabsBackpack = "/static/media/sauce-backpack-1200x1500.0a0b85a3.jpg",
  SauceLabsBikeLight = "/static/media/bike-light-1200x1500.37c843b0.jpg",
  SauceLabsBoltTShirt = "/static/media/bolt-shirt-1200x1500.c2599ac5.jpg",
  SauceLabsFleeceJacket = "/static/media/sauce-pullover-1200x1500.51d7ffaf.jpg",
  SauceLabsOnesie = "/static/media/red-onesie-1200x1500.2ec615b2.jpg",
  TestAllTheThingsTShirtRed = "/static/media/red-tatt-1200x1500.30dadef4.jpg",
}

export enum ProductsNames {
  SauceLabsBackpack = "Sauce Labs Backpack",
  SauceLabsBikeLight = "Sauce Labs Bike Light",
  SauceLabsBoltTShirt = "Sauce Labs Bolt T-Shirt",
  SauceLabsFleeceJacket = "Sauce Labs Fleece Jacket",
  SauceLabsOnesie = "Sauce Labs Onesie",
  TestAllTheThingsTShirtRed = "Test.allTheThings() T-Shirt (Red)",
}

export enum ProductsCartIds {
  SauceLabsBackpack = "sauce-labs-backpack",
  SauceLabsBikeLight = "sauce-labs-bike-light",
  SauceLabsBoltTShirt = "sauce-labs-bolt-t-shirt",
  SauceLabsFleeceJacket = "sauce-labs-fleece-jacket",
  SauceLabsOnesie = "sauce-labs-onesie",
  TestAllTheThingsTShirtRed = "test.allthethings()-t-shirt-(red)",
}